refactor: resolve circular dependency
